Add sign tooltips that display the sign message
re: OpenRCT2/OpenRCT2#8593

Modifies the code for tooltip displays to show the
string assigned to a banner. It also uses the
color assigned to the sign (by sheer coincidence).

As of right now I do not think that it works for
non-banner signs (such as the 3D landscapes or
scrolling signs etc), but it works for the typical
case of banners, as well as "No entry" banners.