Headless servers are no longer players
This simply changes the server startup code to not
add the server as a "player", and leaves player slot 0
empty. The add user command also ignores player 0 when
searching for an unused slot in this case.

Messages sent from the server using "say" will appear as
"[00:00] message here",
in white without a username, as opposed to the
"[00:00] User: message here" in light blue.

I haven't seen any issues arising from testing this
locally but I'm waiting to see if there's any feedback.

[OpenRCT2/OpenRCT2#5141]